DEV-46158 - Change datasources API permissions to Admin Only
Since grafana 8.5 the permission system changed (RBAC was enabled) and the datasources read APIs were changed to be from admin only to be available to other users.
We want to maintain the access to datasources for admin only in these APIs.
We will expose another API with only relevant datasource data we want available.